requirements.txt
# Core Framework
crewai>=0.28.0
llama-index>=0.9.0
langchain>=0.0.267

# AI Models
openai>=1.3.0
sentence-transformers>=2.2.2

# Web Framework
flask>=2.0.1
flask-cors>=3.0.10

# Database
pymongo>=4.3.3
redis>=4.5.1

# Utilities
python-dotenv>=0.19.0
requests>=2.28.1
pydantic>=1.10.8
tqdm>=4.64.1
loguru>=0.6.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0